Moved loan generators to generator package
